+++ src/main/java/onlyoffice/conditions/IsOfficeFileAttachment.java
@@ -33,6 +33,7 @@ import javax.inject.Inject;
 
 public class IsOfficeFileAttachment implements Condition {
     private boolean forEdit;
+    private boolean form;
     private DocumentManager documentManager;
     private AttachmentUtil attachmentUtil;
 
@@ -44,8 +45,12 @@ public class IsOfficeFileAttachment implements Condition {
 
     public void init(Map<String, String> params) throws PluginParseException {
         forEdit = false;
+        form = false;
         if (params != null && !params.isEmpty() && params.get("forEdit") != null) {
             forEdit = !params.get("forEdit").isEmpty();
+        }
+        if (params != null && !params.isEmpty() && params.get("form") != null) {
+            form = !params.get("form").isEmpty();
         }
     }